Refactor the kernel completion logic
